feat: Implement network utility functions for local IP address retrieval and update server discovery logic
